Add graph generation script.
The generated datas do not make any sense. It will increase the
complexity (and size on disk) of the database (which is the goal, but
you've still be warned). The complexity might be too high, especially
regarding the DbPath triggers.

With PostgreSQL, you might want to increase the checkpoint_segments
parameters to something like 64. See:
https://wiki.postgresql.org/wiki/Tuning_Your_PostgreSQL_Server